1. evaluation.py目前已完成eval功能,但是Writer和輸入config的功能還沒完成 2. 新增work_dirs, 之後模型traing的結果都會存在這裡 3. 新增configs, 裡面存放類似於mmdetection的格式

+++ moodle_zoo/yolov7/inference.py
@@ -125,7 +125,7 @@ class Yolov7(baseInference):
                         w = xyxy[2].cpu().numpy() - x
                         h = xyxy[3].cpu().numpy() - y
 
-                        bbox_list.append(list(map(int, [x, y, w, h])))
+                        bbox_list.append(list(map(float, [x, y, w, h])))
                         class_list.append(cls)
                         score_list.append(conf)
                         polygon_list.append(segments[j])
